+++ src/main/java/cambio/simulator/orchestration/scheduling/RandomScheduler.java
@@ -46,7 +46,7 @@ public class RandomScheduler extends Scheduler{
 
         if (pod != null) {
             Node candidateNote = null;
-            int cpuDemand = pod.getCPUDemand();
+            double cpuDemand = pod.getCPUDemand();
             //Doing the same like FirstFitScheduler but nodes are shuffled.
             List<Node> nodes = new ArrayList<>(cluster.getNodes());
             Collections.shuffle(nodes, random);
